+++ application/views/account/account_profile.php
@@ -38,7 +38,7 @@
     <p>
 	<?php if (isset($account_details->picture) && strlen(trim($account_details->picture)) > 0) : ?>
 	<?php echo showPhoto($account_details->picture); ?> &nbsp;
-	<?php echo anchor('account/account_profile/index/delete', '<i class="glyphicon glyphicon-trash"></i> '.lang('profile_delete_picture'), 'class="btn btn-default"'); ?>
+	<?php echo anchor('account/profile/index/delete', '<i class="glyphicon glyphicon-trash"></i> '.lang('profile_delete_picture'), 'class="btn btn-default"'); ?>
 	<?php else : ?>
 		
 		<div class="accountPicSelect clearfix">